requirements.txt
wheel
setuptools
djongo
django
django-mongoengine
sshtunnel
psycopg2
pandas
requests
leveldb
libais
shapely